data migration and add new field in chat model

+++ app/src/main/java/com/joshtalks/joshskills/core/custom_ui/exo_audio_player/AudioModel.kt
@@ -3,5 +3,5 @@ package com.joshtalks.joshskills.core.custom_ui.exo_audio_player
 data class AudioModel(
     val audioUrl: String,
     val tag: String?,
-    val duration: Long
+    val duration: Int
 )
